Move tob/cox to raid command (#4059)
